added heading config in tailwind css and added testing of components
