+++ packages/ui-kit/lib/assets/css/tailwind.css
@@ -1,7 +1,15 @@
 @import 'tailwindcss';
-
+@import url('https://fonts.googleapis.com/css2?family=Nunito:wght@400;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Geist+Mono&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Raleway:wght@300;400;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Figtree:wght@400;700&display=swap');
 
 :root {
+  --font-nunito: 'Nunito', sans-serif;
+  --font-geist-mono: 'Geist Mono', monospace;
+  --font-raleway: 'Raleway', sans-serif;
+  --font-figtree: 'Figtree', sans-serif;
+
   --text-2xs: 10px;
   --text-xs: 12px;
   --text-sm: 14px;
@@ -133,52 +141,60 @@
 }
 
 @layer base {
-  h1 { 
+  h1 {
     font-size: var(--text-4xl);
     font-weight: var(--font-weight-h1);
     line-height: 100%;
     letter-spacing: -0.96px;
+
     @media (min-width: 768px) {
       letter-spacing: -1.28px;
     }
   }
-  h2 { 
+
+  h2 {
     font-size: var(--text-3xl);
     font-weight: var(--font-weight-important);
     line-height: 100%;
     letter-spacing: -0.64px;
+
     @media (min-width: 768px) {
       letter-spacing: -0.96px;
     }
   }
-  h3 { 
+
+  h3 {
     font-size: var(--text-2xl);
     font-weight: var(--font-weight-h3);
     line-height: 110%;
- 
+
   }
-  h4 { 
+
+  h4 {
     font-size: var(--text-xl);
     font-weight: var(--font-weight-important);
     line-height: 120%;
   }
-  h5 { 
+
+  h5 {
     font-size: var(--text-lg);
     font-weight: var(--font-weight-h5);
-   line-height: 120%;
+    line-height: 120%;
   }
-  h6 { 
+
+  h6 {
     font-size: var(--text-md);
     font-weight: var(--font-weight-important);
-   line-height: 120%;
+    line-height: 120%;
   }
 }
+
 .theme-just-do-add {
   --radius-small: 6px;
   --radius-medium: 8px;
   --radius-big: 12px;
   --radius-huge: 16px;
-  
+
   --color-card-fill: #1C1917;
   --color-card-transparent: rgba(28, 25, 23, 0);
   --color-card-stroke: #292524;
@@ -273,10 +289,10 @@
   --color-slider-progress-stroke: var(--color-base-brand-400);
 
   --font-weight-important: 700;
-  --font-weight-normal:400;
-  --font-weight-h1:var(--font-weight-important);
-  --font-weight-h3:var(--font-weight-important);
-  --font-weight-h5:var(--font-weight-important);
+  --font-weight-normal: 400;
+  --font-weight-h1: var(--font-weight-important);
+  --font-weight-h3: var(--font-weight-important);
+  --font-weight-h5: var(--font-weight-important);
   font-family: var(--font-nunito);
 }
 
@@ -394,11 +410,11 @@
   --color-slider-progress-fill: var(--color-base-brand-500);
   --color-slider-progress-stroke: var(--color-base-brand-400);
 
-  --font-weight-important:700;
-  --font-weight-normal:400;
-  --font-weight-h1:var(--font-weight-important);
-  --font-weight-h3:var(--font-weight-important);
-  --font-weight-h5:var(--font-weight-important);
+  --font-weight-important: 700;
+  --font-weight-normal: 400;
+  --font-weight-h1: var(--font-weight-important);
+  --font-weight-h3: var(--font-weight-important);
+  --font-weight-h5: var(--font-weight-important);
   font-family: var(--font-geist-mono);
 }
 
@@ -515,12 +531,12 @@
   --color-slider-knob-stroke: var(--color-base-brand-300);
   --color-slider-progress-fill: var(--color-base-brand-500);
   --color-slider-progress-stroke: var(--color-base-brand-400);
- --font-weight-important:700;
-  --font-weight-normal:400;
+  --font-weight-important: 700;
+  --font-weight-normal: 400;
 
-  --font-weight-h1:300;
-  --font-weight-h3:400;
-  --font-weight-h5:400;
+  --font-weight-h1: 300;
+  --font-weight-h3: 400;
+  --font-weight-h5: 400;
   font-family: var(--font-raleway);
 }
 
@@ -633,10 +649,10 @@
   --color-slider-knob-stroke: var(--color-base-brand-300);
   --color-slider-progress-fill: var(--color-base-brand-500);
   --color-slider-progress-stroke: var(--color-base-brand-400);
- --font-weight-important:700;
-  --font-weight-normal:400;
-  --font-weight-h1:var(--font-weight-important);
-  --font-weight-h3:var(--font-weight-important);
-  --font-weight-h5:var(--font-weight-important);
+  --font-weight-important: 700;
+  --font-weight-normal: 400;
+  --font-weight-h1: var(--font-weight-important);
+  --font-weight-h3: var(--font-weight-important);
+  --font-weight-h5: var(--font-weight-important);
   font-family: var(--font-figtree);
 }
