+++ packages/ui-kit/lib/components/Tabs/TabContent.tsx
@@ -1,6 +1,6 @@
 
 import React from 'react';
-import { useTabContext } from './TabContext';
+import { useTabContext } from './tabContext';
 import { cn } from '../../utils/style-utils';
