+++ packages/ui-kit/lib/components/dateInput.tsx
@@ -1,5 +1,5 @@
 import React, { useRef } from 'react';
-import { IconButton } from '../iconButton';
+import { IconButton } from './iconButton';
 import { Calendar } from 'lucide-react';
 
 export interface DateInputProps {
