+++ packages/ui-kit/lib/components/languageSelector.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { X } from 'lucide-react';
-import { LanguageSelectorProps } from './types';
-import { Button } from '../button';
+import { LanguageSelectorProps } from './profile/types';
+import { Button } from './button';
 
 const AVAILABLE_LANGUAGES = [
   { name: 'English', code: 'en' },
