--- packages/ui-kit/lib/components/professionalCard/types.ts
@@ -1,15 +1,0 @@
-export interface SkillItemProps {
-    name: string;
-  }
-  
-  export interface InputFieldProps {
-    label: string;
-    value: string;
-    optional?: boolean;
-    id: string;
-  }
-  
-  export interface FileDisplayProps {
-    fileName: string;
-    fileSize: string;
-  }
