+++ packages/ui-kit/lib/components/profile/profilePicture.tsx
