+++ packages/ui-kit/lib/components/profile/types.ts
@@ -27,3 +27,19 @@ export interface LanguageSelectorProps {
   selectedLanguages: LanguageProps[];
   onChange: (languages: LanguageProps[]) => void;
 }
+
+export interface SkillItemProps {
+  name: string;
+}
+
+export interface InputFieldProps {
+  label: string;
+  value: string;
+  optional?: boolean;
+  id: string;
+}
+
+export interface FileDisplayProps {
+  fileName: string;
+  fileSize: string;
+}
